Admin CRUD
Update administrator management to allow CRUD support for admins
