+++ src/views/partials/admin/admins/delete.php
@@ -1,3 +1,5 @@
+<?php $this->layout('layouts/default'); ?>
+
 <div class="row">
     <div class="span12">
         <h1>Treasure <small>Delete Treasure</small></h1>
