+++ src/views/partials/admin/admins/edit.php
@@ -1,3 +1,4 @@
+<?php $this->layout('layouts/default'); ?>
 <div class="row">
     <div class="span12">
         <h1>Administrators <small>Edit Administrator</small></h1>
